requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
WTForms==3.2.1
openpyxl==3.1.5
reportlab==4.2.5
Werkzeug==3.1.3
itsdangerous==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
gunicorn==21.2.0
psycopg2-binary==2.9.9
